Reworked mappers tests
