+++ test/UserService.Business.UnitTests/CreateUserCommandTests.cs
@@ -3,6 +3,7 @@ using FluentValidation.Results;
 using LT.DigitalOffice.UserService.Business.Interfaces;
 using LT.DigitalOffice.UserService.Data.Interfaces;
 using LT.DigitalOffice.UserService.Mappers.Interfaces;
+using LT.DigitalOffice.UserService.Mappers.RequestsMappers.Interfaces;
 using LT.DigitalOffice.UserService.Models.Db;
 using LT.DigitalOffice.UserService.Models.Dto;
 using Moq;
@@ -18,8 +19,8 @@ namespace LT.DigitalOffice.UserService.Business.UnitTests
         private Mock<IUserRepository> userRepositoryMock;
         private Mock<IValidator<UserRequest>> validatorMock;
         private Mock<ValidationResult> validationResultIsValidMock;
-        private Mock<IMapper<UserRequest, DbUser>> mapperUserMock;
-        private Mock<IMapper<UserRequest, DbUserCredentials>> mapperUserCredentialsMock;
+        private Mock<IUserRequestMapper> mapperUserMock;
+        private Mock<IUserCredentialsRequestMapper> mapperUserCredentialsMock;
 
         private Guid userId;
         private ValidationResult validationResultError;
@@ -71,8 +72,8 @@ namespace LT.DigitalOffice.UserService.Business.UnitTests
         {
             userRepositoryMock = new Mock<IUserRepository>();
 
-            mapperUserMock = new Mock<IMapper<UserRequest, DbUser>>();
-            mapperUserCredentialsMock = new Mock<IMapper<UserRequest, DbUserCredentials>>();
+            mapperUserMock = new Mock<IUserRequestMapper>();
+            mapperUserCredentialsMock = new Mock<IUserCredentialsRequestMapper>();
 
             validatorMock = new Mock<IValidator<UserRequest>>();
