+++ test/UserService.Business.UnitTests/GetUserByIdCommandTests.cs
@@ -1,7 +1,7 @@
 using LT.DigitalOffice.Kernel.UnitTestLibrary;
 using LT.DigitalOffice.UserService.Business.Interfaces;
 using LT.DigitalOffice.UserService.Data.Interfaces;
-using LT.DigitalOffice.UserService.Mappers.Interfaces;
+using LT.DigitalOffice.UserService.Mappers.ResponsesMappers.Interfaces;
 using LT.DigitalOffice.UserService.Models.Db;
 using LT.DigitalOffice.UserService.Models.Dto;
 using Moq;
@@ -14,7 +14,7 @@ namespace LT.DigitalOffice.UserService.Business.UnitTests
     {
         private IGetUserByIdCommand getUserInfoByIdCommand;
         private Mock<IUserRepository> repositoryMock;
-        private Mock<IMapper<DbUser, User>> mapperMock;
+        private Mock<IUserResponseMapper> mapperMock;
 
         private Guid userId;
         private User user;
@@ -24,7 +24,7 @@ namespace LT.DigitalOffice.UserService.Business.UnitTests
         public void SetUp()
         {
             repositoryMock = new Mock<IUserRepository>();
-            mapperMock = new Mock<IMapper<DbUser, User>>();
+            mapperMock = new Mock<IUserResponseMapper>();
             getUserInfoByIdCommand = new GetUserByIdCommand(repositoryMock.Object, mapperMock.Object);
 
             userId = Guid.NewGuid();
