+++ test/UserService.Mappers.UnitTests/UserMapperTests.cs
@@ -1,22 +1,16 @@
-using LT.DigitalOffice.Broker.Responses;
+using LT.DigitalOffice.Kernel.Exceptions;
 using LT.DigitalOffice.Kernel.UnitTestLibrary;
-using LT.DigitalOffice.UserService.Mappers.Interfaces;
-using LT.DigitalOffice.UserService.Mappers.UnitTests.Utils;
+using LT.DigitalOffice.UserService.Mappers.RequestsMappers;
 using LT.DigitalOffice.UserService.Models.Db;
 using LT.DigitalOffice.UserService.Models.Dto;
 using NUnit.Framework;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using LT.DigitalOffice.Kernel.Exceptions;
-using LT.DigitalOffice.UserService.UserCredentials.Admin;
 
 namespace LT.DigitalOffice.UserService.Mappers.UnitTests
 {
     public class UserMapperTests
     {
-        private IMapper<DbUser, User> mapper;
-        private IMapper<UserRequest, DbUser> mapperEditUserRequest;
+        private UserRequestMapper userRequestMapper;
 
         private const string Message = "smth";
         private const string FirstName = "Ivan";
@@ -39,82 +33,8 @@ namespace LT.DigitalOffice.UserService.Mappers.UnitTests
         [SetUp]
         public void SetUp()
         {
-            mapper = new UserMapper();
-            mapperEditUserRequest = new UserMapper();
-
-            userId = Guid.NewGuid();
-            achievementId = Guid.NewGuid();
-            certificateFileId = Guid.NewGuid();
-            pictureFileId = Guid.NewGuid();
-            avatarFileId = Guid.NewGuid();
-
-            achievement = new DbAchievement
-            {
-                Id = achievementId,
-                Message = Message,
-                PictureFileId = pictureFileId
-            };
-
-            dbUserAchievement = new DbUserAchievement
-            {
-                Achievement = achievement,
-                AchievementId = achievementId,
-                User = dbUser,
-                Time = DateTime.Now,
-                UserId = userId
-            };
-
-            dbUserCertificateFile = new DbUserCertificateFile
-            {
-                CertificateId = certificateFileId,
-                User = dbUser,
-                UserId = userId
-            };
-
-            dbUser = new DbUser
-            {
-                AchievementsIds = new List<DbUserAchievement> { dbUserAchievement },
-                AvatarFileId = avatarFileId,
-                FirstName = FirstName,
-                Id = userId,
-                IsActive = IsActive,
-                IsAdmin = IsAdmin,
-                LastName = LastName,
-                Status = Status,
-                CertificatesFilesIds = new List<DbUserCertificateFile> { dbUserCertificateFile }
-            };
+            userRequestMapper = new UserRequestMapper();
         }
-
-        #region IMapper<DbUser, User>
-
-        [Test]
-        public void ShouldThrowBadRequestExceptionWhenDbUserIsNull()
-        {
-            Assert.Throws<BadRequestException>(() => mapper.Map(null));
-        }
-
-        [Test]
-        public void ShouldReturnUserModelWhenMappingValidDbUser()
-        {
-            var resultUserModel = mapper.Map(dbUser);
-
-            Assert.IsNotNull(resultUserModel);
-            Assert.IsInstanceOf<User>(resultUserModel);
-
-            Assert.AreEqual(achievementId, resultUserModel.AchievementsIds.ToList()[0].Id);
-            Assert.AreEqual(Message, resultUserModel.AchievementsIds.ToList()[0].Message);
-            Assert.AreEqual(pictureFileId, resultUserModel.AchievementsIds.ToList()[0].PictureFileId);
-
-            Assert.AreEqual(certificateFileId, resultUserModel.CertificatesIds.ToList()[0]);
-            Assert.AreEqual(userId, resultUserModel.Id);
-            Assert.AreEqual(FirstName, resultUserModel.FirstName);
-            Assert.AreEqual(LastName, resultUserModel.LastName);
-            Assert.IsNull(resultUserModel.MiddleName);
-            Assert.AreEqual(Status, resultUserModel.Status);
-            Assert.AreEqual(avatarFileId, resultUserModel.AvatarFileId);
-            Assert.AreEqual(IsAdmin, resultUserModel.IsAdmin);
-        }
-        #endregion
 
         #region EditUserRequest to DbUser
         [Test]
@@ -135,7 +55,7 @@ namespace LT.DigitalOffice.UserService.Mappers.UnitTests
                 AvatarFileId = Guid.NewGuid()
             };
 
-            var result = mapperEditUserRequest.Map(request);
+            var result = userRequestMapper.Map(request);
 
             var user = new DbUser()
             {
@@ -158,7 +78,7 @@ namespace LT.DigitalOffice.UserService.Mappers.UnitTests
         {
             UserRequest request = null;
 
-            Assert.Throws<BadRequestException>(() => mapperEditUserRequest.Map(request));
+            Assert.Throws<BadRequestException>(() => userRequestMapper.Map(request));
         }
         #endregion
     }
